Room struct and random generation
I made a struct named "Room" which contains the room grid 2d vector, as well as a flood 2d vector. I think this will make it easier for me when doing AI logic. The flood grid will update whenever the player moves, so the enemies always know the fastest route to the player. The enemies will merely need to compare Enemy.pos to Room.flood to see how close they are to the player.